Add missing index to subtitle suffix Refactor more util methods for files import
