+++ media_manager/torrent/utils.py
@@ -1,11 +1,8 @@
 import hashlib
 import logging
-import mimetypes
-from pathlib import Path
 
 import bencoder
 import libtorrent
-import patoolib
 import requests
 from pathvalidate import sanitize_filename
 from requests.exceptions import InvalidSchema
@@ -13,99 +10,8 @@ from requests.exceptions import InvalidSchema
 from media_manager.config import MediaManagerConfig
 from media_manager.indexer.schemas import IndexerQueryResult
 from media_manager.indexer.utils import follow_redirects_to_final_torrent_url
-from media_manager.torrent.schemas import Torrent
 
 log = logging.getLogger(__name__)
-
-
-def list_files_recursively(path: Path = Path()) -> list[Path]:
-    files = list(path.glob("**/*"))
-    log.debug(f"Found {len(files)} entries via glob")
-    valid_files = []
-    for x in files:
-        if x.is_dir():
-            log.debug(f"'{x}' is a directory")
-        elif x.is_symlink():
-            log.debug(f"'{x}' is a symlink")
-        else:
-            valid_files.append(x)
-    log.debug(f"Returning {len(valid_files)} files after filtering")
-    return valid_files
-
-
-def extract_archives(files: list) -> None:
-    archive_types = {
-        "application/zip",
-        "application/x-zip-compressedapplication/x-compressed",
-        "application/vnd.rar",
-        "application/x-7z-compressed",
-        "application/x-freearc",
-        "application/x-bzip",
-        "application/x-bzip2",
-        "application/gzip",
-        "application/x-gzip",
-        "application/x-tar",
-    }
-    for file in files:
-        file_type = mimetypes.guess_type(file)
-        log.debug(f"File: {file}, Size: {file.stat().st_size} bytes, Type: {file_type}")
-
-        if file_type[0] in archive_types:
-            log.info(
-                f"File {file} is a compressed file, extracting it into directory {file.parent}"
-            )
-            try:
-                patoolib.extract_archive(str(file), outdir=str(file.parent))
-            except patoolib.util.PatoolError:
-                log.exception(f"Failed to extract archive {file}")
-
-
-def get_torrent_filepath(torrent: Torrent) -> Path:
-    return MediaManagerConfig().misc.torrent_directory / torrent.title
-
-
-def get_files_for_import(
-    torrent: Torrent | None = None, directory: Path | None = None
-) -> tuple[list[Path], list[Path], list[Path]]:
-    """
-    Extracts all files from the torrent download directory, including extracting archives.
-    Returns a tuple containing: seperated video files, subtitle files, and all files found in the torrent directory.
-    """
-    if torrent:
-        log.info(f"Importing torrent {torrent}")
-        search_directory = get_torrent_filepath(torrent=torrent)
-    elif directory:
-        log.info(f"Importing files from directory {directory}")
-        search_directory = directory
-    else:
-        msg = "Either torrent or directory must be provided."
-        raise ValueError(msg)
-
-    all_files: list[Path] = list_files_recursively(path=search_directory)
-    log.debug(f"Found {len(all_files)} files downloaded by the torrent")
-    extract_archives(all_files)
-    all_files = list_files_recursively(path=search_directory)
-
-    video_files: list[Path] = []
-    subtitle_files: list[Path] = []
-    for file in all_files:
-        file_type, _ = mimetypes.guess_type(str(file))
-        if file_type is not None:
-            if file_type.startswith("video"):
-                video_files.append(file)
-                log.debug(f"File is a video, it will be imported: {file}")
-            elif file_type.startswith("text") and Path(file).suffix == ".srt":
-                subtitle_files.append(file)
-                log.debug(f"File is a subtitle, it will be imported: {file}")
-            else:
-                log.debug(
-                    f"File is neither a video nor a subtitle, will not be imported: {file}"
-                )
-
-    log.info(
-        f"Found {len(all_files)} files ({len(video_files)} video files, {len(subtitle_files)} subtitle files) for further processing."
-    )
-    return video_files, subtitle_files, all_files
 
 
 def get_torrent_hash(torrent: IndexerQueryResult) -> str:
