+++ media_manager/tv/router.py
@@ -13,7 +13,7 @@ from media_manager.metadataProvider.dependencies import metadata_provider_dep
 from media_manager.metadataProvider.schemas import MetaDataProviderSearchResult
 from media_manager.schemas import MediaImportSuggestion
 from media_manager.torrent.schemas import Torrent
-from media_manager.utils.files import get_importable_media_directories
+from media_manager.utils.file_handler import get_importable_media_directories
 from media_manager.tv.dependencies import (
     season_dep,
     show_dep,
